requirements.txt
APScheduler==3.6.3
python-dotenv==1.1.0
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.3
